취미입력 API 로직 구현

+++ SeSACFriends/SeSACFriends/API/Form.swift
@@ -27,7 +27,7 @@ struct updateMypageForm: Codable {
 
 struct hobbyForm {
     var type: Int
-    var region: Double
+    var region: Int
     var long: Double
     var lat: Double
     var hf: [String]
